requirements.txt
fastapi
uvicorn
voyageai
python-dotenv
Jinja2
scikit-learn
python-multipart